*** DB UPDATE ***  removed mass/thrust ship effect to correct ab/mwd calculations (supporting code not complete for this calculation) renamed misspelling of "turrent" to "turret" updated Bubble::GetSpawnID() to Bubble::GetBeltID() for proper functional naming updated version and added civilian ai version info to server startup display updated effects processing for setting correct charge location updated npc ai - belt rats wander around center of belt when no targets in sight range - belt rats no longer drop targetters when ourside of sight range update effect processing order

+++ src/eve-server/system/cosmicMgrs/SpawnMgr.h
@@ -72,7 +72,7 @@ public:
     bool IsTimerStarted()                               { return m_mainTimer.Enabled(); }
 
 protected:
-    bool _FindSpawnForBubble(uint32 itemID);
+    bool _FindSpawnForBubble(uint16 itemID);
     void PrepSpawn(SystemBubble* pSysBubble, uint32 regionID, double secRating);
     void MakeSpawn(SystemBubble* pSysBubble, uint32 factionID, uint8 type, uint8 subtype);
     void ReSpawn(SystemBubble* pSysBubble, SpawnEntry& spawnEntry);
@@ -88,7 +88,7 @@ protected:
     typedef std::vector<RatSpawnClass> RatSpawnClassVec;
     typedef std::vector<RatFactionGroups> RatFactionGroupsVec;
     typedef std::map<uint8, uint32> RatFactionGroupsMap;    //map to enable 'find'  shipClass is key
-    typedef std::unordered_multimap<uint32, SpawnEntry> SpawnEntryDef;    //bubbleID is key
+    typedef std::unordered_multimap<uint16, SpawnEntry> SpawnEntryDef;    //bubbleID is key
     //typedef std::vector<uint32, SystemSpawnGroup> SystemSpawnGroupVec;  //systemID is key  *unused at this time*
 
 private:
